refactor(app_user): extract _SliverTabBarDelegate to own part file (#2146)
Splits event_detail_content.dart (510 → 463 lines) by moving
_SliverTabBarDelegate to event_detail_tab_bar_delegate.dart.
Part-of structure preserved — no public API change.

Constraint: File uses part/part-of — new file must declare part-of same library
Confidence: high
Scope-risk: narrow
